% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbl-commute-traffic-data.R
\name{get_buses}
\alias{get_buses}
\title{Get Busses for a given location}
\usage{
get_buses(lon = 24.7777, lat = 60.17746,
  time = as.POSIXct("2016-12-07 15:00", format = "\%Y-\%m-\%d \%H:\%M"))
}
\arguments{
\item{lon}{numeric containing current longitude coordinate in format WGS86}

\item{lat}{numeric containing current latitude coordinate in format WGS86}
}
\description{
Returns IDs for the bus stop and expected times for the expected five (default)
busses. The function returns both timestamps for the bus arrivals as
well as minutes.
}
\details{
This function was built for an API to serve the HBL Commute App
}

