% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbl-commute-traffic-data.R
\docType{data}
\name{get_closest_stops}
\alias{get_closest_stops}
\title{Data for every Buss, Train, Tram and Ferry Service}
\format{An object of class \code{"data.frame"}.}
\source{
\href{http://example.com}{Example example}

\href{http://example.com}{Example example}

\href{http://example.com}{Example example}

\href{http://example.com}{Example example}
}
\usage{
data(services)
}
\description{
Bla bla bla bla
Bla bla bla bla
Bla bla bla bla
Bla bla bla bla
}
\examples{
data(services)
lines <- services$line
Schedule Data for every Buss, Train, Tram and Ferry Stop

Bla bla bla bla
Bla bla bla bla
Bla bla bla bla
Bla bla bla bla

}
\keyword{datasets}

